refactor(entity): 패키징 변경
