+++ app/src/main/java/woowacourse/omok/data/db/entity/User.kt
@@ -1,3 +1,3 @@
-package woowacourse.omok.data.db
+package woowacourse.omok.data.db.entity
 
 data class User(val int: Int, val name: String)
